Mori Dream Spaces

This package is a about $\mathbb{C}^*$-surfaces. However, some algorithms here also work for a more general class of varieties admitting a certain embedding into a toric variety, called Mori dream spaces. To allow for the treatment of these varieties in the future, this package provides the abstract MoriDreamSpace type, whose only implementations currently are CStarSurface and ToricSurface.

CStarSurfaces.MoriDreamSpace — Type

MoriDreamSpace

Julia type for Mori Dream Spaces.

Subtypes of MoriDreamSpace should at least implement the following methods:

canonical_toric_ambient(X :: MoriDreamSpace)

cox_ring_relations(X :: MoriDreamSpace)

CStarSurfaces.MoriDreamSpaceUnion — Type

MoriDreamSpaceUnion

The union of MoriDreamSpace and Oscar's NormalToricVarietyType.

CStarSurfaces.canonical_toric_ambient — Function

canonical_toric_ambient(X :: MoriDreamSpace)

Construct the canonical toric ambient variety of a Mori Dream Space.

This function should be implemented by all subtypes of MoriDreamSpace

CStarSurfaces.cox_ring_relations — Function

cox_ring_relations(X :: MoriDreamSpace)

Return the list of relations in the Cox Ring of a Mori Dream Space. Here, a relation is a RingElem whose parent is the Cox Ring of the canonical toric ambient variety.

This function should be implemented by all subtypes of MoriDreamSpace.

Oscar.cox_ring — Method

cox_ring(X :: MoriDreamSpace)

Compute the Cox Ring of a Mori Dream Space.

CStarSurfaces.is_toric — Method

is_toric(X :: MoriDreamSpace)

Check whether a Mori Dream Space is a toric variety.

Oscar.rays — Method

rays(X :: MoriDreamSpace)

Return the rays of the canonical toric ambient variety of a Mori Dream Space.

Oscar.nrays — Method

nrays(X :: MoriDreamSpace)

Return the number of rays of the canonical toric ambient variety of a Mori Dream Space.

Oscar.maximal_cones — Method

maximal_cones(X :: MoriDreamSpace)

Return the maximal cones of the fan of the canonical toric ambient variety of a Mori Dream Space.

CStarSurfaces.maximal_cones_indices — Method

maximal_cones_indices(X :: MoriDreamSpaceUnion)

Return the list of maximal cones of a Mori Dream Space, where each cone is represented as a list of integers that are the indices of its generating rays in the fan.