Surfaces with torus action

The two main Julia types in this packages are CStarSurface and ToricSurface. Some functionality also works for MoriDreamSpace's, an abstract type of which CStarSurface and ToricSurface are subtypes.

Julia types

CStarSurfaces.CStarSurfaceCase — Type

CStarSurfaceCase

The abstract supertype of the possible possible configurations of source and sink of a $\mathbb{C}^*$-surface.

This type has the four subtypes EE, EP, PE and PP, named after the existence of elliptic fixed points and parabolic fixed point curves in the source and sink of a $\mathbb{C}^*$-surface respectively.

CStarSurfaces.CStarSurface — Type

CStarSurface{T<:CStarSurfaceCase}

A $\mathbb{C}^*$-surface of case T <: CStarSurfaceCase. As a Julia type, it gets modeled by a stuct with fields l, d and case, where l and d are zero-indexed vectors of one-indexed vectors of integers and case is one of the four symbols :ee, :pe, :ep and :pp.

CStarSurfaces.ToricSurface — Type

ToricSurface

A toric surface. As a Julia type, it gets modeled by a struct with a single field vs that stores the primitive generator of the rays of the two-dimensional complete fan describing the toric surface.

CStarSurfaces.SurfaceWithTorusAction — Type

SurfaceWithTorusAction = Union{CStarSurface, ToricSurface}

Julia type for surfaces with a non-trivial torus action.

Constructors

CStarSurfaces.cstar_surface — Method

cstar_surface(ls :: Vector{Vector{Int64}}, ds :: Vector{Vector{Int64}}, case :: Symbol)

Construct a C-star surface from the integral vectors $l_i=(l_{i1}, ..., l_{in_i})$ and $d_i=(d_{i1}, ..., d_{in_i})$ and a given C-star surface case. The parameters ls and ds are given both given as a vector of vectors. They must be of the same length and satisfy gcd(ls[i][j], ds[i][j]) == 1 for all i and j. The parameter case can be one of the four symbols :ee, :pe, :ep, :pp.

Example

The $E_6$ singular cubic.

julia> X = cstar_surface([[3, 1], [3], [2]], [[-2, -1], [1], [1]], :ee)
C-star surface of type (e-e)
julia> gen_matrix(X)
[-3   -1   3   0]
[-3   -1   0   2]
[-2   -1   1   1]

CStarSurfaces.cstar_surface — Method

cstar_surface(P :: ZZMatrix)

Construct a C-star surface from a generator matrix of the correct format. That is, $P$ must be of one of the following forms:

\[\begin{array}{lcclc} \text{(e-e)} & \begin{bmatrix} L \\ d \end{bmatrix} & \qquad & \text{(p-e)} & \begin{bmatrix} L & 0 \\ d & 1 \end{bmatrix} \\ \text{(e-p)} & \begin{bmatrix} L & 0 \\ d & -1 \end{bmatrix} & \qquad & \text{(p-p)} & \begin{bmatrix} L & 0 & 0 \\ d & 1 & -1 \end{bmatrix} \end{array},\]

where for some integral vectors $l_i=(l_{i1}, ..., l_{in_i}) \in \mathbb{Z}^{n_i}_{\geq 0}$ and $d_i=(d_{i1}, ..., d_{in_i}) \in \mathbb{Z}^{n_i}$ with $\gcd(l_{ij}, d_{ij}) = 1$, we have

\[L = \begin{bmatrix} -l_0 & l_1 & \dots & 0 \\ \vdots & & \ddots & 0 \\ -l_0 & 0 & \dots & l_r \end{bmatrix}, \qquad d = \begin{bmatrix} d_0 & \dots & d_r \end{bmatrix}.\]

Example

The $E_6$ singular cubic.

julia> cstar_surface(ZZ[-3 -1 3 0 ; -3 -1 0 2 ; -2 -1 1 1])
C-star surface of type (e-e)

CStarSurfaces.toric_surface — Function

toric_surface(vs :: Vector{Vector{T}})

Construct a toric surface from a list of integral vectors in two-dimensional space.

Example

The 5-th Hirzebruch surface.

julia> toric_surface([[1,0], [0,1], [-1,-5], [0,-1]])
Normal toric surface

toric_surface(P :: ZZMatrix)

Construct a toric surface from an integral matrix, where the columns of the matrix are the rays of the describing fan.

Example

The 5-th Hirzebruch surface.

julia> toric_surface(ZZ[1 0 -1 0 ; 0 1 -17 -1])
Normal toric surface

CStarSurfaces.gen_matrix — Function

gen_matrix(X :: MoriDreamSpaceUnion)

Return the generator matrix of a Mori Dream Space X. The columns of this matrix are the rays of the fan of the canonical toric ambient variety of X.

Example

julia> gen_matrix(cstar_surface([[3, 1], [3], [2]], [[-2, -1], [1], [1]], :ee))
[-3   -1   3   0]
[-3   -1   0   2]
[-2   -1   1   1]

Class group and Picard group

Hecke.class_group — Method

class_group(X :: MoriDreamSpace)

Return the class group of a Mori Dream Space.

Example

julia> class_group(cstar_surface([[1, 1], [2], [2]], [[0, -2], [1], [1]], :ee))
GrpAb: Z/4 x Z

CStarSurfaces.class_group_rank — Function

class_group_rank(X :: MoriDreamSpaceUnion)

Return the rank of the class group of a Mori Dream Space.

Example

julia> class_group_rank(cstar_surface([[1, 1], [2], [2]], [[-3, -4], [1], [1]], :pe))
2

CStarSurfaces.class_group_torsion — Function

class_group_torsion(X :: MoriDreamSpaceUnion)

Return the list of elementary divisors that make up the torsion part of the class group of a Mori Dream Space.

Example

julia> class_group_torsion(cstar_surface([[2, 2], [2], [4]], [[3, -3], [1], [1]], :ee))
2-element Vector{ZZRingElem}:
 2
 6

CStarSurfaces.class_group_torsion_order — Function

class_group_torsion_order(X :: MoriDreamSpaceUnion)

Return the order of the torsion part of the class group of a Mori Dream Space.

Example

julia> class_group_torsion_order(cstar_surface([[2, 2], [2], [4]], [[3, -3], [1], [1]], :ee))
12

CStarSurfaces.local_class_groups — Method

local_class_groups(X :: MoriDreamSpace)

Compute the local class groups of a Mori Dream Space X. The result is a dictionary indexed by the maximal cones of X.

Example

julia> local_class_groups(cstar_surface([[2, 2], [2], [4]], [[3, -3], [1], [1]], :ee))
Dict{Vector{Int64}, GrpAbFinGen} with 3 entries:
  [1, 3, 4] => Abelian group with structure: Z/2 x Z/18
  [2, 3, 4] => Abelian group with structure: Z/2 x Z/6
  [1, 2]    => Abelian group with structure: Z/12

CStarSurfaces.local_class_group — Method

local_class_group(X :: MoriDreamSpace, c :: Vector{Int64})

Compute the local class group of a Mori Dream Space X at a maximal cone c given by a vector of indices.

Example

julia> local_class_group(cstar_surface([[2, 2], [2], [4]], [[3, -3], [1], [1]], :ee), [1, 2])
GrpAb: Z/12

CStarSurfaces.maps_from_class_group_to_local_class_groups — Method

maps_from_class_group_to_local_class_groups(X :: MoriDreamSpace)

Compute the canonical maps from the class group of a Mori Dream Space to its local class groups. The result is a dictionary indexed by the maximal cones of X.

CStarSurfaces.map_from_class_group_to_local_class_group — Method

map_from_class_group_to_local_class_group(X :: MoriDreamSpace, c :: Vector{Int64})

Compute the canonical map from the class group of a Mori Dream Space X to its local class group at a cone c, given by a vector of indices.

Hecke.picard_group — Method

picard_group(X :: MoriDreamSpace)

Return the Picard group of a Mori Dream Space.

Example

julia> picard_group(cstar_surface([[1, 1], [2], [2]], [[0, -2], [1], [1]], :ee))
GrpAb: Z

CStarSurfaces.degree_matrix — Function

degree_matrix(X :: MoriDreamSpaceUnion)

Return the degree matrix of a Mori Dream Space (often denoted as $Q$). The columns of this matrix are the degrees of the generator of the Cox Ring, which are elements of the divisor class group. Note that we write the torsion parts of these elements first (in the upper rows of the degree matrix), and the free part after that (in the lower rows of the degree matrix).

Example

julia> degree_matrix(cstar_surface([[2, 2], [2], [4]], [[3, -3], [1], [1]], :ee))
[0   1   1   0]
[0   4   1   5]
[1   3   4   2]

CStarSurfaces.degree_matrix_free_part — Function

degree_matrix_free_part(X :: MoriDreamSpaceUnion)

The free part of the degree matrix of a Mori Dream Space. By the convention in this package, these are the lower rows of degree_matrix(X).

Example

julia> degree_matrix_free_part(cstar_surface([[2, 2], [2], [4]], [[3, -3], [1], [1]], :ee))
[1   3   4   2]

CStarSurfaces.degree_matrix_torsion_part — Function

degree_matrix_torsion_part(X :: MoriDreamSpaceUnion)

The torsion part of the degree matrix of a Mori Dream Space. By the convention in this package, these are the upper rows of degree_matrix(X).

Example

julia> degree_matrix_torsion_part(cstar_surface([[2, 2], [2], [4]], [[3, -3], [1], [1]], :ee))
[0   1   1   0]
[0   4   1   5]

Oscar.gorenstein_index — Method

gorenstein_index(X :: MoriDreamSpace)

Return the Gorenstein index of a $\mathbb{Q}$-Gorenstein Mori Dream Space.

Example

julia> gorenstein_index(cstar_surface([[1, 1], [11], [5]], [[0, -2], [9], [3]], :ee))
78

CStarSurfaces.local_gorenstein_indices — Function

local_gorenstein_indices(X :: MoriDreamSpaceUnion)

Return the local gorenstein indices of a Mori Dream Space. The result is a dictionary indexed by the maximal cones of X.

Example

julia> local_gorenstein_indices(cstar_surface([[2, 2], [2], [4]], [[3, -3], [1], [1]], :ee))
Dict{Vector{Int64}, ZZRingElem} with 3 entries:
  [1, 3, 4] => 9
  [2, 3, 4] => 3
  [1, 2]    => 2

CStarSurfaces.local_gorenstein_index — Function

local_gorenstein_index(X :: MoriDreamSpaceUnion, c :: Vector{Int64})

Return the local gorenstein index of a Mori Dream Space at a given cone c, which must be given as a vector of ray indices.

Example

julia> local_gorenstein_index(cstar_surface([[2, 2], [2], [4]], [[3, -3], [1], [1]], :ee), [1, 3, 4])
9

Oscar.picard_index — Method

picard_index(X :: MoriDreamSpace)

Return the index of the Picard group in the class group of a Mori Dream Space.

Example

julia> picard_index(cstar_surface([[1, 1], [7], [7]], [[0, -1], [3], [3]], :ee))
42

CStarSurfaces.local_picard_indices — Function

local_picard_indices(X :: MoriDreamSpaceUnion)

Return the local picard indices of a Mori Dream Space. The result is a dictionary indexed by the maximal cones of X.

Example

julia> local_picard_indices(cstar_surface([[2, 2], [2], [4]], [[3, -3], [1], [1]], :ee))
Dict{Vector{Int64}, ZZRingElem} with 3 entries:
  [1, 3, 4] => 36
  [2, 3, 4] => 12
  [1, 2]    => 12

CStarSurfaces.local_picard_index — Function

local_picard_index(X :: MoriDreamSpaceUnion, c :: Vector{Int64})

Return the local picard index of a Mori Dream Space at a given cone c, which must be given as a vector of ray indices.

Example

julia> local_picard_index(cstar_surface([[2, 2], [2], [4]], [[3, -3], [1], [1]], :ee), [1, 3, 4])
36

CStarSurfaces.is_factorial — Function

is_factorial(X :: MoriDreamSpace)

Determine if a Mori Dream Space has at most factorial singularities, i.e. its canonical toric ambient variety is smooth.

Example

julia> is_factorial(cstar_surface([[1, 1], [7], [7]], [[0, -1], [3], [3]], :ee))
false

Divisors

CStarSurfaces.CStarSurfaceDivisor — Type

CStarSurfaceDivisor{T}

A Weil divisor on a $\mathbb{C}^*$-surface of type T <: CStarSurfaceCase.

CStarSurfaces.cstar_surface_divisor — Function

cstar_surface_divisor(X :: CStarSurface, coeffs :: Vector{<:Oscar.IntegerUnion})

Construct a divisor on a C-star surface as a linear combination of the the torus invariant prime divisors.

Example

julia> X = cstar_surface([[3, 1], [3], [2]], [[2, 1], [1], [1]], :ee)
C-star surface of type (e-e)

julia> D = cstar_surface_divisor(X, [0, 1, -1, 3])
CStarSurfaceDivisor{EE}(C-star surface of type (e-e), Torus-invariant, non-prime divisor on a normal toric variety)

julia> coefficients(D)
4-element Vector{ZZRingElem}:
 0
 1
 -1
 3

cstar_surface_divisor(X :: CStarSurface, coeffs :: DoubleVector{<:Oscar.IntegerUnion})

Construct a divisor on a C-star surface as a linear combination of the the torus invariant prime divisors, where the coefficients are given following the double index notation.

Example

julia> X = cstar_surface([[3, 1], [3], [2]], [[2, 1], [1], [1]], :ee)
C-star surface of type (e-e)

julia> D = cstar_surface_divisor(X, DoubleVector([[0, 1], [-1], [3]]))
CStarSurfaceDivisor{EE}(C-star surface of type (e-e), Torus-invariant, non-prime divisor on a normal toric variety)

julia> coefficients(D)
4-element Vector{ZZRingElem}:
 0
 1
 -1
 3

CStarSurfaces.ToricSurfaceDivisor — Type

ToricSurfaceDivisor

A Weil divisor on a toric surface.

CStarSurfaces.toric_surface_divisor — Function

toric_surface_divisor(X :: ToricSurface, coeffs :: Vector{S})

Construct a divisor on a toric surface as a linear combination of the the torus invariant prime divisors.

Example

julia> X = toric_surface([[1,0], [0,1], [-1,-5], [0,-1]])
Normal toric surface

julia> D = toric_surface_divisor(X, [1, 2, -1, 5])
ToricSurfaceDivisor(Normal toric surface, Torus-invariant, non-prime divisor on a normal toric variety)

julia> coefficients(D)
4-element Vector{ZZRingElem}:
 1
 2
 -1
 5

CStarSurfaces.SurfaceWithTorusActionDivisor — Type

SurfaceWithTorusActionDivisor = Union{CStarSurfaceDivisor, ToricSurfaceDivisor}

A Weil divisor on a surface with non-trivial torus action.

Hecke.canonical_divisor — Method

canonical_divisor(X :: MoriDreamSpace)

Return the canonical divisor of a Mori Dream Space.

Example

julia> X = cstar_surface([[1, 1], [2], [2]], [[0, -2], [1], [1]], :ee)
C-star surface of type (e-e)

julia> coefficients(canonical_divisor(X))
4-element Vector{ZZRingElem}:
 -1
 -1
 1
 -1

Oscar.anticanonical_divisor — Method

anticanonical_divisor(X :: MoriDreamSpace)

Return the anticanonical divisor of a Mori Dream Space.

Example

julia> X = cstar_surface([[1, 1], [2], [2]], [[0, -2], [1], [1]], :ee)
C-star surface of type (e-e)

julia> coefficients(anticanonical_divisor(X))
4-element Vector{ZZRingElem}:
 1
 1
 -1
 1

Oscar.canonical_divisor_class — Method

canonical_divisor_class(X :: MoriDreamSpace)

Return the canonical divisor class of a Mori Dream Space.

Example

julia> X = cstar_surface([[1, 1], [2], [2]], [[0, -2], [1], [1]], :ee)
C-star surface of type (e-e)

julia> divisor_class(canonical_divisor_class(X))
Element of
GrpAb: Z/4 x Z
with components [0 -2]

Oscar.anticanonical_divisor_class — Method

anticanonical_divisor_class(X :: MoriDreamSpace)

Return the anticanonical divisor class of a Mori Dream Space.

Example

julia> X = cstar_surface([[1, 1], [2], [2]], [[0, -2], [1], [1]], :ee)
C-star surface of type (e-e)

julia> divisor_class(anticanonical_divisor_class(X))
Element of
GrpAb: Z/4 x Z
with components [0 2]

Base.:* — Method

Base.:*(d1 :: SurfaceWithTorusActionDivisor, d2 :: SurfaceWithTorusActionDivisor)

Return the intersection number of two divisors on a surface with torus action. The divisors have to be defined on the same variety.

Example

julia> X = cstar_surface([[3, 1], [3], [2]], [[-2, -1], [1], [1]], :ee)
C-star surface of type (e-e)

julia> D = cstar_surface_divisor(X, [1, 0, 2, -1])
CStarSurfaceDivisor{EE}(C-star surface of type (e-e), Torus-invariant, non-prime divisor on a normal toric variety)

julia> D * D
4//3

Intersection numbers

CStarSurfaces.intersection_matrix — Function

intersection_matrix(X :: ToricSurface)

Return the matrix of intersection numbers of all toric prime divisors of a toric surface X with each other. The result is a rational n x n matrix, where n = nrays(X) and the (i,j)-th entry is the intersection number of the toric prime divisors associated to the i-th and j-th ray respectively.

Example

julia> intersection_matrix(toric_surface(ZZ[1 0 -1 0 ; 0 1 -17 -1]))
[0    1   0     1]
[1   17   1     0]
[0    1   0     1]
[1    0   1   -17]

intersection_matrix(X :: CStarSurface)

Return the matrix of intersection numbers of all restrictions of toric prime divisors of a C-star surface X with each other. The result is a rational n x n matrix, where n = nrays(X) and the (i,j)-th entry is the intersection number of the toric prime divisors associated to the i-th and j-th ray respectively.

Example

julia> intersection_matrix(cstar_surface([[3, 1], [3], [2]], [[-2, -1], [1], [1]], :ee))
[1//3   1   2//3   1]
[   1   3      2   3]
[2//3   2   4//3   2]
[   1   3      2   3]

CStarSurfaces.anticanonical_self_intersection — Function

anticanonical_self_intersection(X :: SurfaceWithTorusAction)

Return the self intersection number of the anticanonical divisor on a surface with torus action.

Example

julia> anticanonical_self_intersection(cstar_surface([[3, 1], [3], [2]], [[-2, -1], [1], [1]], :ee))
3

Resolution of singularities

CStarSurfaces.canonical_resolution — Function

canonical_resolution(X :: ToricSurface)

Return the canonical resolution of singularities of a toric surface X. The result is a triple (Y, ex_rays, discrepancies) where Y is the smooth toric surface in the resolution of singularities of X, ex_rays contains the rays of the exceptional divisors in the resolution and discrepancies contains their discrepancies.

Example

julia> X = toric_surface(ZZ[1 1 -3 ; 0 4 -7])
Normal toric surface

julia> (Y, ex, discr) = canonical_resolution(X);

julia> gen_matrix(Y)
[1   1   -3   1   1   1   0   -1   -2   -1    0]
[0   4   -7   1   2   3   1   -2   -5   -3   -1]

julia> ex
3-element Vector{Vector{Vector{Int64}}}:
 [[1, 1], [1, 2], [1, 3]]
 [[0, 1], [-1, -2]]
 [[-2, -5], [-1, -3], [0, -1]]

julia> discr
3-element Vector{Vector{Rational{Int64}}}:
 [0//1, 0//1, 0//1]
 [-1//5, -2//5]
 [-1//7, -2//7, -3//7]

canonical_resolution(X :: CStarSurface)

Return the canonical resolution of singularities of a C-star surface surface X. The result is a triple (Y, ex_rays, discrepancies) where Y is the smooth C-star surface in the resolution of singularities of X, ex_rays contains the rays of the exceptional divisors in the resolution and discrepancies contains their discrepancies. ex_rays and discrepancies itself are given as pairs, where the first entry comes from the toric step and the second entry from the tropical step. The first entry is given as a DoubleVector, adhering to the double index notation of C-star surfaces.

Example

Resolution of singularities of the $E_6$ singular cubic surface.

julia> X = cstar_surface([[3, 1], [3], [2]], [[-2, -1], [1], [1]], :ee)
C-star surface of type (e-e)

julia> (Y, (ex_toric, ex_tropic), (discr_toric, discr_tropic)) = canonical_resolution(X);

julia> gen_matrix(Y)
[-3   -1   -1   -2   3   1   2   1   0   0   0   0    0]
[-3   -1   -1   -2   0   0   0   0   2   1   1   0    0]
[-2   -1    0   -1   1   1   1   0   1   1   0   1   -1]

julia> ex_toric
3-element DoubleVector{Vector{Vector{Int64}}} with indices ZeroRange(3):
 [[[1, 0], [2, -1]], [], []]
 [[[1, 1], [2, 1]], [[1, 0]]]
 [[[1, 1]], [[1, 0]]]

julia> ex_tropic
2-element Vector{Vector{Int64}}:
 [0, 1]
 [0, -1]

julia> discr_toric
3-element DoubleVector{Vector{Rational{Int64}}} with indices ZeroRange(3):
 [[0//1, 0//1], [], []]
 [[0//1, 0//1], [1//1]]
 [[0//1], [2//1]]

julia> discr_tropic
2-element Vector{Rational{Int64}}:
 0//1
 4//1

CStarSurfaces.exceptional_rays — Function

exceptional_rays(X :: SurfaceWithTorusAction)

Return the rays associated to the exceptional divisors in the resolution of singularities of a surface with torus action.

CStarSurfaces.discrepancies — Function

discrepancies(X :: SurfaceWithTorusAction)

Return the discrepancies associated to the exceptional divisors in the resolution of singularities of a surface with torus action.

CStarSurfaces.maximal_log_canonicity — Function

maximal_log_canonicity(X :: ToricSurface)

Given a toric surface $X$, return the maximal rational number $\varepsilon$ such that $X$ is $\varepsilon$-log canonical. By definition, this is the minimal discrepancy in the resolution of singularities plus one.

Example

julia> X = toric_surface(ZZ[1 1 -3 ; 0 4 -7])
Normal toric surface

julia> maximal_log_canonicity(X)
4//7

maximal_log_canonicity(X :: CStarSurface)

Given a C-star surface $X$, return the maximal rational number $\varepsilon$ such that $X$ is $\varepsilon$-log canonical. By definition, this is the minimal discrepancy in the resolution of singularities plus one.

Example

julia> maximal_log_canonicity(cstar_surface([[3, 1], [3], [2]], [[-2, -1], [1], [1]], :ee))
1//1

Attributes of $\mathbb{C}^*$-surfaces

CStarSurfaces.canonical_toric_ambient — Method

canonical_toric_ambient(X :: CStarSurface)

Construct the canonical toric ambient of a C-star surface, as an Oscar type.

Example

julia> X = cstar_surface([[3, 1], [3], [2]], [[-2, -1], [1], [1]], :ee)
C-star surface of type (e-e)

julia> Z = canonical_toric_ambient(X)
Normal toric variety

julia> rays(Z)
4-element SubObjectIterator{RayVector{QQFieldElem}}:
 [-1, -1, -2//3]
 [-1, -1, -1]
 [1, 0, 1//3]
 [0, 1, 1//2]

CStarSurfaces.cox_ring_vars — Method

cox_ring_vars(X :: CStarSurface)

Return the variables in the Cox Ring of the canonical toric ambient of a C-star surface, following the double index notation. The result is a tuple, whose first entry is a DoubleVector consisting of the variables T[i][j] and whose second entry is the Vector of variables [S[1], ... S[m]], where m is the number of parabolic fixed point curves.

Example

julia> X = cstar_surface([[1, 1], [2], [2]], [[-3, -4], [1], [1]], :pe)
C-star surface of type (p-e)

julia> (Ts, Ss) = cox_ring_vars(X);

julia> Ts
3-element DoubleVector{MPolyDecRingElem} with indices ZeroRange(3):
 [T[0][1], T[0][2]]
 [T[1][1]]
 [T[2][1]]

julia> Ss
1-element Vector{MPolyDecRingElem{QQFieldElem, QQMPolyRingElem}}:
 S[1]

CStarSurfaces.cox_ring_relations — Method

cox_ring_relations(X :: CStarSurface)

Return the list of relations in the Cox Ring of a C-star surface. Note that these are all trinomials.

Example

julia> X = cstar_surface([[1, 1], [2], [2]], [[-3, -4], [1], [1]], :pe)
C-star surface of type (p-e)

julia> cox_ring_relations(X)
1-element Vector{MPolyDecRingElem{QQFieldElem, QQMPolyRingElem}}:
 T[0][1]*T[0][2] + T[1][1]^2 + T[2][1]^2

CStarSurfaces.has_x_plus — Method

has_x_plus(X :: CStarSurface{<:CStarSurfaceCase})

Checks whether a given C-star surface has an elliptic fixed point in the source, commonly denoted $x^+$.

CStarSurfaces.has_x_minus — Method

has_x_minus(X :: CStarSurface{<:CStarSurfaceCase})

Checks whether a given C-star surface has an elliptic fixed point in the sink, commonly denoted $x^-$.

CStarSurfaces.has_D_plus — Method

has_D_plus(X :: CStarSurface{<:CStarSurfaceCase})

Checks whether a given C-star surface has a parabolic fixed point curve in the source, commonly denoted $D^+$.

CStarSurfaces.has_D_minus — Method

has_D_minus(X :: CStarSurface{<:CStarSurfaceCase})

Checks whether a given C-star surface has a parabolic fixed point curve in the sink, commonly denoted $D^-$.

CStarSurfaces.nblocks — Function

nblocks(X :: CStarSurface)

Returns the number of blocks in the generator matrix of a C-star surface.

CStarSurfaces.block_sizes — Function

block_sizes(X :: CStarSurface)

Returns the sizes of the blocks in the generator matrix of a C-star surface. The result is a zero-indexed vector of type ZeroVector.

CStarSurfaces.number_of_parabolic_fixed_point_curves — Function

number_of_parabolic_fixed_point_curves(X :: CStarSurface)

Returns the number of parabolic fixed point curves of a C-star surface.

CStarSurfaces.slopes — Function

slopes(X :: CStarSurface)

Returns the DoubleVector of slopes of a C-star surface, i.e a DoubleVector with slopes(X)[i][j] = X.d[i][j] // X.l[i][j].

CStarSurfaces.is_intrinsic_quadric — Function

is_intrinsic_quadric(X :: CStarSurface)

Checks whether a C-star surface is an intrinsic quadric, i.e its Cox Ring has a single quadratic relation.

CStarSurfaces.cox_ring_vars — Function

cox_ring_vars(X :: CStarSurface)

Return the variables in the Cox Ring of the canonical toric ambient of a C-star surface, following the double index notation. The result is a tuple, whose first entry is a DoubleVector consisting of the variables T[i][j] and whose second entry is the Vector of variables [S[1], ... S[m]], where m is the number of parabolic fixed point curves.

Example

julia> X = cstar_surface([[1, 1], [2], [2]], [[-3, -4], [1], [1]], :pe)
C-star surface of type (p-e)

julia> (Ts, Ss) = cox_ring_vars(X);

julia> Ts
3-element DoubleVector{MPolyDecRingElem} with indices ZeroRange(3):
 [T[0][1], T[0][2]]
 [T[1][1]]
 [T[2][1]]

julia> Ss
1-element Vector{MPolyDecRingElem{QQFieldElem, QQMPolyRingElem}}:
 S[1]